feat!: azapi 2.0 (#103)
* feat!: azapi v2 update

BREAKING CHANGE: refactored to remove the helper module and use the azapi features directly

* feat: Update deprecated variables and resources

Update the deprecated variables and resources to align with the latest changes in the azapi library. This includes updating the delays and retry configurations for management groups, policy definitions, policy set definitions, role definitions, and subscription placement. The changes ensure compatibility with the new `retries` variable and remove the deprecated `delays` variable.

* fix: temporarily remove retry error for role assignments awaiting https://github.com/Azure/terraform-provider-azapi/pull/588

* fix: role definition id resource id should be tenant level for MG assignments

* feat: update retries and fmt folowwing merge of Azure/terraform-provider-azapi#588

* feat: add example for template architecture definition

* fix: increase policy assignment create timeout to 15 minutes

* fix: add in other mg fields for retries

* fix: revert to derfaults for mg retries

* feat: update to azapi 2.0 beta

* chore: avmfix

* chore: remove dependabot

* docs: make docs